app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
passlib[bcrypt]>=1.7.4
python-jose[cryptography]
apscheduler
transformers
torch
accelerate
sendgrid
twilio
pydantic[email]